@@ -2,58 +2,58 @@ package archives.tater.phantomstun.mixin;
 
 import archives.tater.phantomstun.PhantomStun;
 import archives.tater.phantomstun.Stunnable;
-import net.minecraft.entity.EntityType;
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.damage.DamageSource;
-import net.minecraft.entity.mob.MobEntity;
-import net.minecraft.entity.mob.PhantomEntity;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.entity.projectile.PersistentProjectileEntity;
-import net.minecraft.server.world.ServerWorld;
-import net.minecraft.storage.ReadView;
-import net.minecraft.storage.WriteView;
-import net.minecraft.world.World;
+import net.minecraft.server.level.ServerLevel;
+import net.minecraft.world.damagesource.DamageSource;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.Mob;
+import net.minecraft.world.entity.monster.Phantom;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.entity.projectile.arrow.AbstractArrow;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.storage.ValueInput;
+import net.minecraft.world.level.storage.ValueOutput;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-@Mixin(PhantomEntity.class)
-public abstract class PhantomEntityStunMixin extends MobEntity implements Stunnable {
+@Mixin(Phantom.class)
+public abstract class PhantomEntityStunMixin extends Mob implements Stunnable {
 	@Unique
 	private int phantomstun$stunnedTicks = 0;
 
-	protected PhantomEntityStunMixin(EntityType<? extends MobEntity> entityType, World world) {
+	protected PhantomEntityStunMixin(EntityType<? extends Mob> entityType, Level world) {
 		super(entityType, world);
 	}
 
 	@Inject(
-			method = "writeCustomData",
+			method = "addAdditionalSaveData",
 			at = @At("TAIL")
 	)
-	private void writeStun(WriteView view, CallbackInfo ci) {
+	private void writeStun(ValueOutput view, CallbackInfo ci) {
 		view.putInt("Stunned", phantomstun$stunnedTicks);
 	}
 
 	@Inject(
-			method = "readCustomData",
+			method = "readAdditionalSaveData",
 			at = @At("TAIL")
 	)
-	private void readStun(ReadView view, CallbackInfo ci) {
-		phantomstun$stunnedTicks = view.getInt("Stunned", 0);
+	private void readStun(ValueInput view, CallbackInfo ci) {
+		phantomstun$stunnedTicks = view.getIntOr("Stunned", 0);
 	}
 
 	@Override
-	public boolean damage(ServerWorld serverWorld, DamageSource source, float amount) {
-		if (!super.damage(serverWorld, source, amount)) return false;
+	public boolean hurtServer(ServerLevel serverWorld, DamageSource source, float amount) {
+		if (!super.hurtServer(serverWorld, source, amount)) return false;
 
-		var sourceEntity = source.getSource();
-		if (source.isIn(PhantomStun.ALWAYS_STUN_DAMAGE_TAG)
-				|| (sourceEntity != null && sourceEntity.getType().isIn(PhantomStun.ALWAYS_STUN_ENTITY_TAG))
+		var sourceEntity = source.getDirectEntity();
+		if (source.is(PhantomStun.ALWAYS_STUN_DAMAGE_TAG)
+				|| (sourceEntity != null && sourceEntity.getType().is(PhantomStun.ALWAYS_STUN_ENTITY_TAG))
 				// Players are handled separately
-				|| (sourceEntity instanceof LivingEntity livingEntity && !(source.getAttacker() instanceof PlayerEntity) && source.isIn(PhantomStun.MELEE_STUN_TAG) && livingEntity.getWeaponDisableBlockingForSeconds() > 0)
-				|| (sourceEntity instanceof PersistentProjectileEntity projectile && projectile.isCritical())) {
+				|| (sourceEntity instanceof LivingEntity livingEntity && !(source.getEntity() instanceof Player) && source.is(PhantomStun.MELEE_STUN_TAG) && livingEntity.getSecondsToDisableBlocking() > 0)
+				|| (sourceEntity instanceof AbstractArrow projectile && projectile.isCritArrow())) {
 			phantomstun$setStunned();
 		}
 
