@@ -1,7 +1,7 @@
 package archives.tater.phantomstun.mixin;
 
 import archives.tater.phantomstun.Stunnable;
-import net.minecraft.entity.mob.PhantomEntity;
+import net.minecraft.world.entity.monster.Phantom;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -9,12 +9,12 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
-@Mixin(targets = "net.minecraft.entity.mob.PhantomEntity$FindTargetGoal")
+@Mixin(targets = "net.minecraft.world.entity.monster.Phantom$PhantomAttackPlayerTargetGoal")
 public class FindTargetGoalMixin {
-    @Shadow @Final PhantomEntity field_7319;
+    @Shadow @Final Phantom field_7319;
 
     @Inject(
-            method = "canStart",
+            method = "canUse",
             at = @At("HEAD"),
             cancellable = true)
     private void checkStun(CallbackInfoReturnable<Boolean> cir) {
