@@ -1,7 +1,7 @@
 package archives.tater.phantomstun.mixin;
 
 import archives.tater.phantomstun.Stunnable;
-import net.minecraft.entity.mob.PhantomEntity;
+import net.minecraft.world.entity.monster.Phantom;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -10,21 +10,21 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.Slice;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-@Mixin(targets = "net.minecraft.entity.mob.PhantomEntity$PhantomMoveControl")
+@Mixin(targets = "net.minecraft.world.entity.monster.Phantom$PhantomMoveControl")
 public class PhantomMoveControlMixin {
 
-    @Shadow @Final PhantomEntity field_7330;
+    @Shadow @Final Phantom field_7330;
 
     @Inject(
             method = "tick",
-            at = @At(value = "INVOKE", target = "Lnet/minecraft/entity/mob/PhantomEntity;getYaw()F"),
+            at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/monster/Phantom;getYaw()F"),
             slice = @Slice(
-                    from = @At(value = "INVOKE", target = "Lnet/minecraft/entity/mob/PhantomEntity;setPitch(F)V")
+                    from = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/monster/Phantom;setPitch(F)V")
             ),
             cancellable = true)
     private void checkStun(CallbackInfo ci) {
         if (((Stunnable) field_7330).phantomstun$isStunned()) {
-            field_7330.setVelocity(field_7330.getVelocity().add(0.0, -0.02, 0.0));
+            field_7330.setDeltaMovement(field_7330.getDeltaMovement().add(0.0, -0.02, 0.0));
             ci.cancel();
         }
     }
