@@ -1,7 +1,7 @@
 package archives.tater.phantomstun.mixin;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
-import net.minecraft.world.spawner.PhantomSpawner;
+import net.minecraft.world.level.levelgen.PhantomSpawner;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 
@@ -10,7 +10,7 @@ import static java.lang.Math.max;
 @Mixin(PhantomSpawner.class)
 public class PhantomSpawnerCountMixin {
     @ModifyExpressionValue(
-            method = "spawn",
+            method = "tick",
             at = @At(value = "INVOKE", target = "Lnet/minecraft/world/Difficulty;getId()I")
     )
     private int decreaseSpawnCount(int original) {
